backend/requirements.txt
python-multipart>=0.0.9
transformers>=4.0.0
torch>=2.0.0
numpy>=1.26.0
spacy>=3.7.0
sentence-transformers>=3.0.0
openai>=1.45.0
google-auth>=2.35.0
lightgbm>=4.0.0
xgboost>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
beanie==2.0.0
motor>=3.5.0
pydantic-settings>=2.4.0
email-validator>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
requests>=2.32.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
playwright>=1.46.0
redis>=5.0.0
prometheus-client>=0.20.0
aiosmtplib>=3.0.2
pypdf>=4.3.0
python-docx>=1.1.2
duckdb>=1.1.0
clickhouse-connect>=0.8.12
boto3>=1.35.0
